Multiple Lanes Adding Option
I updated the Traffic Manager Script, so that you can add spawn points (in other words: lanes) and I made the parameters a class which you can also add to each spawn point, so each lane has different parameters of spawning the cars. If you add a parameter to the list but do not assign values it will get the default values from the scripts, and if you have more parameters than spawn points that is not a problem.